ISSUE #3632 - FormSingleSelect becomes FormSelect
