+++ frontend/src/v5/ui/controls/formMultiSelect/multiSelectMenuItem/multiSelectMenuItem.styles.ts
