+++ frontend/src/v5/ui/controls/formSelect/formSearchSelect/formSearchSelect.styles.ts
