+++ frontend/stories/inputs/select/FormSelect.stories.tsx
@@ -15,58 +15,96 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 import { FormSelect } from '@controls/formSelect/formSelect.component';
-import { MenuItem } from '@mui/material';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 import { useForm } from 'react-hook-form';
+import ClearIcon from '@assets/icons/controls/clear_circle.svg';
+import ChevronIcon from '@assets/icons/chevron.svg';
+import CalendarIcon from '@assets/icons/outlined/calendar-outlined.svg';
+import PrintIcon from '@assets/icons/print.svg';
+import { SelectMenuItem } from '@controls/formSelect/formSelectMenutItem.styles';
 import { FormContainer } from '../FormInput.styles';
+import { IconContainer, ItemContainer } from './FormSelect.styles';
 
 export default {
-	title: 'Inputs/Select/FormSelect',
+	title: 'Inputs/Select/FormSingleSelect',
 	argTypes: {
 		label: {
 			type: 'string',
 		},
 		defaultValue: {
-			type: 'string',
-		},
-		values: {
-			control: 'array',
-		},
-		formError: {
-			type: 'string',
+			type: 'number',
 		},
 		disabled: {
 			type: 'boolean',
 		},
+		search: {
+			type: 'boolean',
+		},
+		formError: {
+			type: 'string',
+		},
 	},
 	component: FormSelect,
-	parameters: { controls: { exclude: ['control', 'ref'] } },
+	parameters: { controls: { exclude: ['control', 'ref', 'onItemClick', 'itemIsSelected', 'selectedOptionsTooltip', 'renderValueTooltip'] } },
 } as ComponentMeta<typeof FormSelect>;
 
-const Controlled: ComponentStory<typeof FormSelect> = ({ formError, values, ...args }: any) => {
+const Controlled: ComponentStory<typeof FormSelect> = ({ formError, ...args }: any) => {
 	const { control } = useForm({ mode: 'onChange' });
 
 	return (
 		<FormContainer>
 			<FormSelect
-				name="select"
+				name="multiselect"
 				control={control}
-				{...args}
+				renderValue={(children) => (
+					children?.length > 1 ? (
+						<ItemContainer>
+							<IconContainer>
+								{children[0]}
+							</IconContainer>
+							{children[1]}
+						</ItemContainer>
+					) : (children)
+				)}
 				formError={formError ? { message: formError } : null}
+				{...args}
 			>
-				{values.map((value) => (
-					<MenuItem value={value} key={value} style={{ padding: '8px 14px' }}>
-						{value}
-					</MenuItem>
-				))}
+				<SelectMenuItem value={1} key={1}>
+					<IconContainer>
+						<CalendarIcon />
+					</IconContainer>
+					calendar icon
+				</SelectMenuItem>
+				<SelectMenuItem value={2} key={2}>
+					<IconContainer>
+						<ClearIcon />
+					</IconContainer>
+					clear icon
+				</SelectMenuItem>
+				<SelectMenuItem value={3} key={3} disabled>
+					<IconContainer>
+						<ChevronIcon />
+					</IconContainer>
+					chevron icon
+				</SelectMenuItem>
+				<SelectMenuItem value={4} key={4}>
+					<IconContainer>
+						<PrintIcon />
+					</IconContainer>
+					print icon
+				</SelectMenuItem>
 			</FormSelect>
 		</FormContainer>
 	);
 };
 
 export const ControlledFormSelect = Controlled.bind({});
-// More on args: https://storybook.js.org/docs/react/writing-stories/args
 ControlledFormSelect.args = {
-	label: 'Controlled Select input',
-	values: ['value 1', 'value 2', 'value 3', 'Longer value 4'],
+	label: 'Controlled Search Single Select input',
+};
+
+export const ControlledFormSelectWithDefault = Controlled.bind({});
+ControlledFormSelectWithDefault.args = {
+	label: 'Controlled Search Single Select input',
+	defaultValue: 3,
 };
